uipcp: enrollment: Neighbor and NeighFlow dynamically allocated
This is necessary since they are non-trivial containers, with fds,
state, CDAP connections, etc, so we don't want to trigger
default constructors, copy constructors and destructors.

Also call the uipcp fini ops before finalizing the event loop: This
prevents the finalization function to empty timer lists and event
lists, so that the user can free them manually. In general, we
need it since the fini operation may need the event-loop services.