rina-normal: limit rtxq size
When rtxq reaches its maximum size, the normal IPCP responds with
backpressure, in the same way it would respond if the cwq was
full. In this way the rtxq does not grow indefinitely. It also
makes sense to block the sender application when it is consuming
too much sender (rtxq) buffer.