uipcp: normal: Neighbor::remote_sync_rib: introduce hardwired limit
When RIB subtrees becomes too big, they cannot be encoded in a
single CDAP messages. A proper solution to this problem would
require SDU fragementation. As a temporary solution, however,
we split long lists of objects when sending the RIB to a remote
peer. The limit is hardwired for now, but can be computed exactly
once we know the length of the serialized list of objects.