kernel: separate flow_put and flow_del functionalities
A list flows_putq is used to link flows for which flow_put() is
going to be called at expiration time (never bound, postponed
flow removal).
Now flow_put only drops the reference counter, and when it reaches
zero it only unlinks the flow from the hash tables and puts it
into the flows_removeq list.
A work_queue is used to drain the flows_removeq list and call
flow_del() in process context, which does all the necessary
remove operations.