kernel: normal: remove tail padding introduced by lower layer on receive
On the receive side, it may happen that pci->pdu_len < rb->len,
that is rb length (including current PCI header) is larger than the
length stored in pci->pdu_len. As an example, this happens when
one of the lower layers is a shim-eth, and pci->pdu_len is less
than 46 bytes (the minimum Ethernet payload length).
This patch detects this case and removes the tail padding.