kernel: efcp: advertise rcv_lwe_priv in the nack_ack_seq_num field
This is important for retransmission to work correctly, otherwise
the sender can enter a loop where it tries to retransmit the
same PDUs again and again because the receiver application is
not draining the RX queue (and so increasing rcv_lwe).

On the other end, rcv_lwe (and not rcv_lwe_priv) must be published
to the sender for flow control to work properly.