user, kernel: add support for rxhdroom
If local DIF stacking is asymmetric, it may be necessary for a
shim IPCP to reserve some headroom when allocating receive
packets. Receive headroom will be needed for packets that are
forwarded along a transmit chain of IPCP longer than the receive
chain. This happens when the tree of local IPCPs is not balanced.