uipcp: container: fix subtle bug in uipcp_loop() signaling
We can't use eventfd to send commands, as those commands are summed
up by the kernel. The reader can then occasionally read the sum of two
or more commands, so it's hard to decode what commands were received.
To solve the issue we simply pass the command out of band.