uipcps: PolicyBuilder(): replace std::list with std::vector
The default argument initialization for lists, e.g.

    std::list<std::string> ps = {}

has been observed to cause bugs (segfaults, infinite loops) on
some older GCC versions (libstdc++.so.6.0.19). We therefore
replace the std::list with a std::vector, which does not show
the same issue. An alternative fix would be to use the extended
initialization syntax:

    std::list<std::string> ps = std::list<std::string>()

but this option has been rejected because more redunant.