uipcp: introduce policy dependencies
Two new functions have been introduced:
- policy_register(), which registers a policy, and takes an optional
list of dependencies in the form <component, policy_name> pairs as the
last argument.
- policy_register_group(), which additionally creates cyclical dependencies
  between all members of the group.

policy_register() also checks for coflicting dependencies (i.e. multiple
policies for the same component) in the dependency list, and fails if
that is the case.

policy_mod() now calls policy_deps_get() to traverse the dependency tree
of the desired policy to build a list of <component, policy> pairs.
Again, checks for conflicting (indirect) dependencies are done. If this
succeeds, the whole list of policies is activated in reverse order, i.e.
the policy originally requested is activated last.